move errors to errors.ts
